Fixes:    Debug output fixes. Details:  On Windows, sending a char value of < 0 to                  isprint() will cause an abort. ;-)
[git-p4: depot-paths = "//depot/qt/main/": change = 243315]

+++ src/network/qnativesocketengine_win.cpp
@@ -90,7 +90,7 @@ static QByteArray qt_prettyDebug(const char *data, int len, int maxLength)
     QByteArray out;
     for (int i = 0; i < len; ++i) {
         char c = data[i];
-        if (isprint(c)) {
+        if (isprint(int(uchar(c)))) {
             out += c;
         } else switch (c) {
         case '\n': out += "\\n"; break;
